requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
cryptography==41.0.7
pycryptodome==3.19.0
psutil==5.9.6
matplotlib
reportlab==4.0.7
Werkzeug==3.0.1
python-dotenv==1.0.0
argon2-cffi==23.1.0
